+++ eNMS/services/miscellaneous/ansible_playbook.py
@@ -57,7 +57,12 @@ class AnsiblePlaybookService(Service):
         full_command = " ".join(command + arguments)
         if password:
             full_command = full_command.replace(password, "*" * 10)
-        run.log("info", f"Sending Ansible playbook: {full_command}", device)
+        run.log(
+            "info",
+            f"Sending Ansible playbook: {full_command}",
+            device,
+            logger="security",
+        )
         try:
             result = check_output(
                 command + arguments, cwd=app.path / "files" / "playbooks"